ref(command): move to forge package and add missing header
