+++ proxy/src/main/java/com/velocitypowered/proxy/protocol/packet/brigadier/forge/ModIdArgumentProperty.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.velocitypowered.proxy.protocol.packet.brigadier;
+package com.velocitypowered.proxy.protocol.packet.brigadier.forge;
 
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.arguments.ArgumentType;
